server/swagger_server/response_code/updates_controller.py
import connexion
import six

from swagger_server.models.facility_update_patch import FacilityUpdatePatch  # noqa: E501
from swagger_server.models.facility_update_post import FacilityUpdatePost  # noqa: E501
from swagger_server.models.status200_ok_no_content import Status200OkNoContent  # noqa: E501
from swagger_server.models.status400_bad_request import Status400BadRequest  # noqa: E501
from swagger_server.models.status401_unauthorized import Status401Unauthorized  # noqa: E501
from swagger_server.models.status403_forbidden import Status403Forbidden  # noqa: E501
from swagger_server.models.status404_not_found import Status404NotFound  # noqa: E501
from swagger_server.models.status500_internal_server_error import Status500InternalServerError  # noqa: E501
from swagger_server.models.updates import Updates  # noqa: E501
from swagger_server import util
from swagger_server.response_code import updates_controller as rc


def updates_get():  # noqa: E501
    """Facility Updates (placeholder)

    Facility Updates (placeholder) # noqa: E501


    :rtype: Updates
    """
    return 'do some magic!'


def updates_post(body=None):  # noqa: E501
    """Facility Updates (placeholder)

    Facility Updates (placeholder) # noqa: E501

    :param body: Create a facility update
    :type body: dict | bytes

    :rtype: Updates
    """
    return 'do some magic!'


def updates_uuid_delete(uuid):  # noqa: E501
    """Facility Updates (placeholder)

    Facility Updates (placeholder) # noqa: E501

    :param uuid: universally unique identifier
    :type uuid: str

    :rtype: Status200OkNoContent
    """
    return 'do some magic!'


def updates_uuid_get(uuid):  # noqa: E501
    """Facility Updates (placeholder)

    Facility Updates (placeholder) # noqa: E501

    :param uuid: universally unique identifier
    :type uuid: str

    :rtype: Updates
    """
    return 'do some magic!'


def updates_uuid_patch(uuid, body=None):  # noqa: E501
    """Facility Updates (placeholder)

    Facility Updates (placeholder) # noqa: E501

    :param uuid: universally unique identifier
    :type uuid: str
    :param body: Update an existing facility update
    :type body: dict | bytes

    :rtype: Status200OkNoContent
    """
    return 'do some magic!'
